# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Gestione Dispensa Public API
  version: 2.0.0
  description: Canonical browser-facing API exposed by the Gateway. Domain services remain internal network boundaries.
servers:
  - url: https://<host>:8443/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /privacy/consents:
    get: {tags: [Privacy], operationId: getConsents, responses: {'200': {$ref: '#/components/responses/Data'}}}
    put: {tags: [Privacy], operationId: updateConsents, requestBody: {$ref: '#/components/requestBodies/Json'}, responses: {'200': {$ref: '#/components/responses/Data'}}}
  /privacy/export:
    post: {tags: [Privacy], operationId: requestExport, parameters: [{$ref: '#/components/parameters/IdempotencyKey'}], responses: {'202': {$ref: '#/components/responses/Data'}}}
  /privacy/erasure:
    post: {tags: [Privacy], operationId: requestErasure, parameters: [{$ref: '#/components/parameters/IdempotencyKey'}], responses: {'202': {$ref: '#/components/responses/Data'}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
...
    IdempotencyKey: {name: X-Idempotency-Key, in: header, required: false, schema: {type: string, minLength: 8, maxLength: 256}}
  requestBodies:
    Json:
      required: true
      content:
        application/json:
          schema: {type: object, additionalProperties: true}
  responses:
    Data:
      description: Successful response
      content:
        application/json:
          schema: {$ref: '#/components/schemas/DataEnvelope'}
...
  schemas:
    DataEnvelope:
      type: object
      required: [data]
      properties:
        data: {}
        meta: {type: object, additionalProperties: true}
...